# packages/knowledge-control-plane/docs/api/generic-v1.openapi.json
{
  "components": {
    "schemas": {
      "ArtifactContentResponse": {
        "additionalProperties": false,
        "properties": {
          "artifact": {
            "$ref": "#/components/schemas/ArtifactSummary"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "next_offset": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "schema_version": {
            "default": "generic_api/v1",
            "title": "Schema Version",
            "type": "string"
          },
          "truncated": {
            "default": false,
            "title": "Truncated",
            "type": "boolean"
          }
        },
        "required": [
          "artifact",
          "content"
        ],
        "title": "ArtifactContentResponse",
        "type": "object"
      },
...
      "CapabilitiesResponse": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "schema_version": {
            "default": "generic_api/v1",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "capabilities"
        ],
        "title": "CapabilitiesResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Knowledge Profile execution, self-contained AISL publication bundle creation, and typed Producer artifacts.",
    "title": "Knowledge Control Plane Knowledge Execution API",
    "version": "generic_api/v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_api_v1_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource state conflict"
          },
          "422": {
            "description": "Request validation failed"
          },
          "500": {
            "description": "Internal backend error"
          },
          "502": {
            "description": "Invalid upstream response"
          },
          "503": {
            "description": "Required capability unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        },
        "summary": "Get Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/content": {
      "get": {
        "operationId": "get_artifact_content_api_v1_artifacts__artifact_id__content_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500000,
              "maximum": 5000000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource state conflict"
          },
          "422": {
            "description": "Request validation failed"
          },
          "500": {
            "description": "Internal backend error"
          },
          "502": {
            "description": "Invalid upstream response"
          },
          "503": {
            "description": "Required capability unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        },
        "summary": "Get Artifact Content",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/artifacts/{artifact_id}/download": {
      "get": {
        "operationId": "download_artifact_api_v1_artifacts__artifact_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {}
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource state conflict"
          },
          "422": {
            "description": "Request validation failed"
          },
          "500": {
            "description": "Internal backend error"
          },
          "502": {
            "description": "Invalid upstream response"
          },
          "503": {
            "description": "Required capability unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "artifacts"
        ]
      }
    },
    "/api/v1/capabilities": {
      "get": {
        "operationId": "get_capabilities_api_v1_capabilities_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource state conflict"
          },
          "422": {
            "description": "Request validation failed"
          },
          "500": {
            "description": "Internal backend error"
          },
          "502": {
            "description": "Invalid upstream response"
          },
          "503": {
            "description": "Required capability unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        },
        "summary": "Get Capabilities",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/v1/configuration": {
      "get": {
        "operationId": "get_configuration_api_v1_configuration_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource state conflict"
          },
          "422": {
            "description": "Request validation failed"
          },
          "500": {
            "description": "Internal backend error"
          },
          "502": {
            "description": "Invalid upstream response"
          },
          "503": {
            "description": "Required capability unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        },
        "summary": "Get Configuration",
        "tags": [
          "configuration"
        ]
      },
      "put": {
        "operationId": "update_configuration_api_v1_configuration_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource state conflict"
          },
          "422": {
            "description": "Request validation failed"
          },
          "500": {
            "description": "Internal backend error"
          },
          "502": {
            "description": "Invalid upstream response"
          },
          "503": {
            "description": "Required capability unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        },
        "summary": "Update Configuration",
        "tags": [
          "configuration"
        ]
      }
    },
    "/api/v1/configuration/validate": {
      "post": {
        "operationId": "validate_configuration_api_v1_configuration_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigurationValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationValidationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          },
          "409": {
            "description": "Resource state conflict"
          },
          "422": {
            "description": "Request validation failed"
          },
          "500": {
            "description": "Internal backend error"
          },
          "502": {
            "description": "Invalid upstream response"
          },
          "503": {
            "description": "Required capability unavailable"
          },
          "504": {
            "description": "Upstream timeout"
          }
        },
        "summary": "Validate Configuration",
        "tags": [
          "configuration"
        ]
      }
    },
...
  },
...
}